backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
qrcode[pil]==7.4.2
Pillow==10.2.0
httpx==0.27.0
pytest==8.0.0
pytest-asyncio==0.23.5
